requirements.txt
click==8.0.1
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.2
Flask==2.0.1
Flask-Bootstrap==3.3.7.1
Flask-MySQLdb==0.2.0
Flask-WTF==0.15.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysqlclient==2.0.3
passlib==1.7.4
pycryptodome==3.10.1
visitor==0.1.3
Werkzeug==2.0.1
WTForms==2.3.3
mysql-connector-python~=8.0.26
requests~=2.26.0